Alter Catch* and stop using internal packages
The internal packages are convenient but insert
panic dumping into function calls.

Instead the caller can Catch and then dump the panic.
Similarly, for normal errors we want to capture a stack
and wrap the errors with that.
At the top-level of main or the go routine a decision
will be made how to print out stack information.