Try takes handlers directly and wraps errors
By default, wrap errors with a stack trace

Try, but not TryX takes handlers directly.
Not sure if this is better because it is inconsistent
with TryX.
However, with the prior version it was possible to
not add handlers to `Try` and thus not execute code.

For TryX an assignment should take place.
If no handler is given, Go will normally give a loud error message
because the type of the assigned variable will be incorrect.